+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.css';
-import WelcomeModal from "./components/welcome/welcomeModal";
+import WelcomeModal from "./components/welcome/WelcomeModal";
 
 function App() {
   return (
